requirements.txt
fastapi[standard]
uvicorn
sqlalchemy
pydantic
asyncpg
alembic
python-dotenv
langgraph
crewai
pytest
pytest-asyncio
httpx
psycopg2-binary
